# server/docs/swagger.yaml
basePath: /
definitions:
...
  model.Record:
    properties:
      account:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      domain:
        type: string
      id:
        description: 主键ID
        type: integer
      line:
        type: integer
      record_id:
        type: string
      rr:
        type: string
      status:
        type: integer
      type:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      value:
        type: string
    type: object
  model.Records:
    properties:
      records:
        items:
          $ref: '#/definitions/model.Record'
        type: array
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  request.RecordListSearch:
    properties:
      account:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      domain:
        type: string
      id:
        description: 主键ID
        type: integer
      line:
        type: integer
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      record_id:
        type: string
      rr:
        type: string
      status:
        type: integer
      type:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /recordList/UpdateRecords:
    put:
      consumes:
      - application/json
      parameters:
      - description: 批量更新Record
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Records'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量更新Record
      tags:
      - Record
  /recordList/createRecord:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建Record
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Record'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建Record
      tags:
      - Record
  /recordList/deleteRecord:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除Record
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Record'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除Record
      tags:
      - Record
  /recordList/deleteRecordByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除Record
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除Record
      tags:
      - Record
  /recordList/findRecord:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询Record
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Record'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询Record
      tags:
      - Record
  /recordList/flushRecords:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 强制刷新域名解析
      tags:
      - Record
  /recordList/getRecordList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页获取record列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RecordListSearch'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取Record列表
      tags:
      - Record
  /recordList/updateRecord:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新Record
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Record'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新Record
      tags:
      - Record
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"